Kerix_scraper/requirements.txt
bs4
pymongo
selenium
selenium-stealth
webdriver-manager
tqdm
twocaptcha
python-dotenv